send DTMF via hidden call (#19)
* intercom: add hidden call

* intercom: wait until all DTMF keys where sent

A call is initiated with "Subject: hidden" SIP header that informs local and
peer module intercom that the call should be hidden for other application
modules. When the call becomes established the TX audio stream is used to
send a DTMF key sequence to the peer. As soon as all keys are transmitted the
call is terminated.

* intercom: add NULL pointer checks